Replace old generator with jinja2 generator
- Replace the templates
- Rename jinja2 generator to the original generators name
