+++ python/generator_utils.py
@@ -4,11 +4,6 @@ Module holding some generator utility functions
 """
 from __future__ import unicode_literals, absolute_import, print_function
 
-from copy import deepcopy
-
-from podio_templates import declarations, implementations
-
-
 def _get_namespace_class(full_type):
   """Get the namespace and the unqualified classname from the full type. Raise a
   DefinitionError if a nested namespace is found"""
@@ -23,210 +18,11 @@ def _get_namespace_class(full_type):
 
   return "", full_type
 
-
-def demangle_classname(classname):
-  """Get the namespace, classname as well as opening and closing statements (for
-  the namespace) from the full classname"""
-  namespace, rawclassname = _get_namespace_class(classname)
-  if namespace:
-    namespace_open = "namespace {} {{".format(namespace)
-    namespace_close = "}} // namespace {}".format(namespace)
-  else:
-    namespace_open, namespace_close = "", ""
-
-  return namespace, rawclassname, namespace_open, namespace_close
-
-
 def _prefix_name(name, prefix):
   """Prefix the name and capitalize the first letter if the prefix is not empty"""
   if prefix:
     return prefix + name[0].upper() + name[1:]
   return name
-
-
-def _format(string, **replacements):
-  """Do 'format'-like replacements of the passed key-value pairs in strings that
-  also contain other curly-braces"""
-  res_string = string
-  for key, repl in replacements.items():
-    res_string = res_string.replace('{' + key + '}', repl)
-  return res_string
-
-
-def get_extra_code(classname, definition):
-  """Get the extra code from a datatype definition"""
-  extra_code = {'decl': "", 'const_decl': "", 'impl': "", 'const_impl': "", 'includes': set()}
-
-  extra = definition.get('ExtraCode', {})
-  const_extra = definition.get('ConstExtraCode', {})
-  if not extra and not const_extra:
-    return extra_code
-
-  extra_code['decl'] = _format(extra.get('declaration', ''), name=classname)
-  extra_code['impl'] = _format(extra.get('implementation', ''), name=classname)
-
-  extra_code['const_decl'] = _format(const_extra.get('declaration', ''), name='Const' + classname)
-  extra_code['decl'] += '\n' + _format(const_extra.get('declaration', ''), name=classname)
-
-  extra_code['const_impl'] = _format(const_extra.get('implementation', ''), name='Const' + classname)
-  extra_code['impl'] += '\n' + _format(const_extra.get('implementation', ''), name=classname)
-
-  # For the moment join the includes for both cases
-  extra_code['includes'].update(extra.get('includes', '').split('\n'))
-  extra_code['includes'].update(const_extra.get('includes', '').split('\n'))
-
-  return extra_code
-
-
-def _get_format_pattern_func(default_replacements):
-  """Get a format function that has some default replacements and can additionally
-  take some more replacements that override the defaults"""
-  def _format_pattern(pattern_str, **replacements):
-    """format the given string updating the default_replacements as necessary"""
-    # make the arguments to this override the defaults if need be
-    repls = deepcopy(default_replacements)
-    repls.update(replacements)
-    return pattern_str.format(**repls)
-
-  return _format_pattern
-
-
-def get_fmt_func(**replacements):
-  """Wraper that works with kwargs"""
-  return _get_format_pattern_func(replacements)
-
-
-def _make_return_dict(getter_decls, getter_impls, setter_decls, setter_impls, const_getter_impls):
-  """Put all the generation code into a dict to have some structure. Arguments can
-  either be a list of strings or just plain strings"""
-  return {
-      'decl': {
-          'get': ''.join(getter_decls),
-          'set': ''.join(setter_decls)
-          },
-      'impl': {
-          'get': ''.join(getter_impls),
-          'set': ''.join(setter_impls),
-          'const_get': ''.join(const_getter_impls)
-          }
-      }
-
-
-def generate_get_set_member(member, classname, get_syntax, components=None):
-  """Create the getters and setters for members of the class if components are
-  passed the members of the components will also be exposed"""
-  default_replacements = {
-      'type': member.full_type,
-      'name': member.name,
-      'description': member.description,
-      'classname': classname
-      }
-  _format_pattern = _get_format_pattern_func(default_replacements)
-  getname, setname = member.getter_setter_names(get_syntax)
-
-  getter_decls = []
-  getter_impls = []
-  setter_decls = []
-  setter_impls = []
-  const_getter_impls = []
-
-  getter_decls.append(_format_pattern(declarations["member_getter"], fname=getname))
-  getter_impls.append(_format_pattern(implementations["member_getter"], fname=getname))
-  const_getter_impls.append(_format_pattern(implementations["const_member_getter"], fname=getname))
-
-  if member.is_builtin or member.is_array:
-    setter_decls.append(_format_pattern(declarations["member_builtin_setter"], fname=setname))
-    setter_impls.append(_format_pattern(implementations["member_builtin_setter"], fname=setname))
-
-    if member.is_array:
-      # common replacements
-      repls = {'type': member.array_type, 'fname': getname}
-      getter_decls.append(_format_pattern(declarations["array_member_getter"], **repls))
-      getter_impls.append(_format_pattern(implementations["array_member_getter"], **repls))
-      const_getter_impls.append(_format_pattern(implementations["const_array_member_getter"], **repls))
-
-      repls['fname'] = setname
-      setter_decls.append(_format_pattern(declarations["array_builtin_setter"], **repls))
-      setter_impls.append(_format_pattern(implementations["array_builtin_setter"], **repls))
-
-  else:
-    # To be compatible with previously generated code, the get/set naming scheme
-    # is not used for getting the non-const references
-    setter_decls.append(_format_pattern(declarations["member_class_refsetter"], fname=member.name))
-    setter_impls.append(_format_pattern(implementations["member_class_refsetter"], fname=member.name))
-
-    setter_decls.append(_format_pattern(declarations["member_class_setter"], fname=setname))
-    setter_impls.append(_format_pattern(implementations["member_class_setter"], fname=setname))
-
-    if components is not None:
-      for sub_member in components[member.full_type]['Members']:
-        getname, setname = sub_member.getter_setter_names(get_syntax)
-
-        repls = {'type': sub_member.full_type, 'name': sub_member.name,
-                 'fname': getname, 'compname': member.name}
-        getter_decls.append(_format_pattern(declarations["pod_member_getter"], **repls))
-        getter_impls.append(_format_pattern(implementations["pod_member_getter"], **repls))
-        const_getter_impls.append(_format_pattern(implementations["const_pod_member_getter"], **repls))
-
-        repls['fname'] = setname
-        if sub_member.is_builtin:
-          setter_decls.append(_format_pattern(declarations["pod_member_builtin_setter"], **repls))
-          setter_impls.append(_format_pattern(implementations["pod_member_builtin_setter"], **repls))
-
-        else:
-          setter_decls.append(_format_pattern(declarations["pod_member_class_refsetter"], **repls))
-          setter_impls.append(_format_pattern(implementations["pod_member_class_refsetter"], **repls))
-
-          setter_decls.append(_format_pattern(declarations["pod_member_class_setter"], **repls))
-          setter_impls.append(_format_pattern(implementations["pod_member_class_setter"], **repls))
-
-  # give this thing a minimum amount structure
-  return _make_return_dict(getter_decls, getter_impls, setter_decls, setter_impls, const_getter_impls)
-
-
-def generate_get_set_relation(relation, classname, get_syntax):
-  """Create getters and setters for 'OneToOneRelations'"""
-  namespace, cls, _, _ = demangle_classname(relation.full_type)
-  default_replacements = {
-      'name': relation.name,
-      'description': relation.description,
-      'classname': classname,
-      'namespace': namespace,
-      'type': cls
-      }
-  _format_pattern = _get_format_pattern_func(default_replacements)
-
-  getname, setname = relation.getter_setter_names(get_syntax)
-
-  getter_decl = _format_pattern(declarations['one_rel_getter'], fname=getname)
-  getter_impl = _format_pattern(implementations['one_rel_getter'], fname=getname)
-  const_getter_impl = _format_pattern(implementations['const_one_rel_getter'], fname=getname)
-
-  setter_decl = _format_pattern(declarations['one_rel_setter'], fname=setname)
-  setter_impl = _format_pattern(implementations['one_rel_setter'], fname=setname)
-
-  return _make_return_dict(getter_decl, getter_impl, setter_decl, setter_impl, const_getter_impl)
-
-
-def constructor_destructor_collection(relations, references, vectormembers):
-  """Generate the constructor and destructor bodys for a collection"""
-  constructor = []
-  destructor = []
-
-  if relations or references:
-    destructor.append('  for (auto& pointer : m_refCollections) { if (pointer) delete pointer; }')
-
-  for relation in relations + references:
-    destructor.append(implementations["destroy_relations"].format(name=relation.name))
-    constructor.append('  m_refCollections.push_back(new std::vector<podio::ObjectID>());')
-
-  for vecmem in vectormembers:
-    destructor.append('  if (m_vec_{name}) delete m_vec_{name};'.format(name=vecmem.name))
-    repls = {'type': vecmem.full_type, 'name': vecmem.name}
-    constructor.append('  m_vecmem_info.emplace_back("{type}", &m_vec_{name});'.format(**repls))
-    constructor.append('  m_vec_{name} = new std::vector<{type}>();'.format(**repls))
-
-  return '\n'.join(constructor), '\n'.join(destructor)
 
 
 class DefinitionError(Exception):
