+++ python/templates/Collection.cc.jinja2
