+++ python/templates/Collection.h.jinja2
