--- python/templates/CollectionPrepareForWritingVecMember.cc.template
@@ -1,4 +1,0 @@
-   (*m_data)[i].${name}_begin = ${name}_index;
-   (*m_data)[i].${name}_end += ${name}_index;
-   ${name}_index = (*m_data)[i].${name}_end;
-   for(const auto& it : (*m_vecs_${name}[i])) { m_vec_${name}->push_back(it); }
