--- python/templates/CollectionSetReferences.cc.template
@@ -1,12 +1,0 @@
-  for(unsigned int i=0, size=m_refCollections[${counter}]->size(); i!=size; ++i) {
-    const auto id = (*m_refCollections[${counter}])[i];
-    if (id.index != podio::ObjectID::invalid) {
-      CollectionBase* coll = nullptr;
-      collectionProvider->get(id.collectionID, coll);
-      ${class}Collection* tmp_coll = static_cast<${class}Collection*>(coll);
-      const auto tmp = (*tmp_coll)[id.index];
-      m_rel_${name}->emplace_back(tmp);
-    } else {
-      m_rel_${name}->emplace_back(nullptr);
-    }
-  }
