--- python/templates/CollectionSetSingleReference.cc.template
@@ -1,11 +1,0 @@
-  for(unsigned int i = 0, size = m_entries.size(); i != size; ++i) {
-    const auto id = (*m_refCollections[${counter}])[i];
-    if (id.index != podio::ObjectID::invalid) {
-      CollectionBase* coll = nullptr;
-      collectionProvider->get(id.collectionID, coll);
-      ${class}Collection* tmp_coll = static_cast<${class}Collection*>(coll);
-      m_entries[i]->m_${name} = new Const${rawclass}((*tmp_coll)[id.index]);
-    } else {
-      m_entries[i]->m_${name} = nullptr;
-    }
-  }
