+++ python/templates/Component.h.jinja2
