--- python/templates/Component.h.template
@@ -1,19 +1,0 @@
-#ifndef ${PACKAGE_NAME}_${name}_H
-#define ${PACKAGE_NAME}_${name}_H
-
-$includes
-
-#include <ostream>
-
-${namespace_open}
-class $name {
-public:
-$members
-$extracode_declarations
-};
-
-${ostreamComponents}
-
-${namespace_close}
-
-#endif
