+++ python/templates/ConstObject.cc.jinja2
