--- python/templates/ConstObject.cc.template
@@ -1,71 +1,0 @@
-// datamodel specific includes
-#include "${incfolder}${name}.h"
-#include "${incfolder}${name}Const.h"
-#include "${incfolder}${name}Obj.h"
-#include "${incfolder}${name}Data.h"
-#include "${incfolder}${name}Collection.h"
-
-$includes_cc
-#include <ostream>
-
-${namespace_open}
-
-Const${name}::Const${name}() : m_obj(new ${name}Obj()) {
- m_obj->acquire();
-}
-
-${constructor_implementation}
-
-Const${name}::Const${name}(const Const${name}& other) : m_obj(other.m_obj) {
-  m_obj->acquire();
-}
-
-Const${name}& Const${name}::operator=(const Const${name}& other) {
-  if (m_obj) m_obj->release();
-  m_obj = other.m_obj;
-  return *this;
-}
-
-Const${name}::Const${name}(${name}Obj* obj) : m_obj(obj) {
-  if(m_obj)
-    m_obj->acquire();
-}
-
-Const${name} Const${name}::clone() const {
-  return {new ${name}Obj(*m_obj)};
-}
-
-Const${name}::~Const${name}(){
-  if (m_obj) m_obj->release();
-}
-
-$getters
-$relations
-$constextracode
-bool  Const${name}::isAvailable() const {
-  if (m_obj) {
-    return true;
-  }
-  return false;
-}
-
-const podio::ObjectID Const${name}::getObjectID() const {
-  if (m_obj) {
-    return m_obj->id;
-  }
-  return podio::ObjectID{podio::ObjectID::invalid, podio::ObjectID::invalid};
-}
-
-bool Const${name}::operator==(const ${name}& other) const {
-     return (m_obj == other.m_obj);
-}
-
-//bool operator< (const ${name}& p1, const ${name}& p2 ) {
-//  if( p1.m_containerID == p2.m_containerID ) {
-//    return p1.m_index < p2.m_index;
-//  } else {
-//    return p1.m_containerID < p2.m_containerID;
-//  }
-//}
-
-${namespace_close}
