--- python/templates/ConstObject.constructor.cc.template
@@ -1,4 +1,0 @@
-Const${name}::Const${name}(${signature}) : m_obj(new ${name}Obj()){
- m_obj->acquire();
- ${constructor}
-}
