+++ python/templates/ConstObject.h.jinja2
