--- python/templates/Data.h.template
@@ -1,18 +1,0 @@
-#ifndef ${PACKAGE_NAME}_${name}DATA_H
-#define ${PACKAGE_NAME}_${name}DATA_H
-
-$includes
-
-${namespace_open}
-/** @class ${name}Data
- *  $description
- *  @author: $author
- */
-
-class ${name}Data {
-public:
-$members
-};
-${namespace_close}
-
-#endif
