+++ python/templates/Obj.cc.jinja2
