--- python/templates/Obj.cc.template
@@ -1,25 +1,0 @@
-#include "${incfolder}${name}Obj.h"
-$includes_cc
-
-${namespace_open}
-${name}Obj::${name}Obj() :
-    ObjBase{{podio::ObjectID::untracked, podio::ObjectID::untracked}, 0}, data()${initialize_relations}
-{ }
-
-${name}Obj::${name}Obj(const podio::ObjectID id, ${name}Data data) :
-    ObjBase{id, 0}, data(data)
-{ }
-
-${name}Obj::${name}Obj(const ${name}Obj& other) :
-    ObjBase{{podio::ObjectID::untracked, podio::ObjectID::untracked}, 0}, data(other.data)${deepcopy_relations}
-{
-${set_deepcopy_relations}
-}
-
-${name}Obj::~${name}Obj() {
-  if (id.index == podio::ObjectID::untracked) {
-${delete_relations}
-  }
-${delete_singlerelations}
-}
-${namespace_close}
