+++ python/templates/Obj.h.jinja2
