--- python/templates/Obj.h.template
@@ -1,42 +1,0 @@
-#ifndef ${PACKAGE_NAME}_${name}OBJ_H
-#define ${PACKAGE_NAME}_${name}OBJ_H
-
-// data model specific includes
-#include "${incfolder}${name}Data.h"
-
-$includes
-
-#include "podio/ObjBase.h"
-
-// std includes
-#include <atomic>
-
-// forward declarations
-$forward_declarations
-
-${namespace_open}
-
-class ${name};
-class Const${name};
-
-
-class ${name}Obj : public podio::ObjBase {
-public:
-  /// constructor
-  ${name}Obj();
-  /// copy constructor (does a deep-copy of relation containers)
-  ${name}Obj(const ${name}Obj&);
-  /// constructor from ObjectID and ${name}Data
-  /// does not initialize the internal relation containers
-  ${name}Obj(const podio::ObjectID id, ${name}Data data);
-  virtual ~${name}Obj();
-
-public:
-  ${name}Data data;
-${relations}
-
-};
-${namespace_close}
-
-
-#endif
