+++ python/templates/Object.cc.jinja2
