--- python/templates/Object.cc.template
@@ -1,76 +1,0 @@
-// datamodel specific includes
-#include "${incfolder}${name}.h"
-#include "${incfolder}${name}Const.h"
-#include "${incfolder}${name}Obj.h"
-#include "${incfolder}${name}Data.h"
-#include "${incfolder}${name}Collection.h"
-
-$includes_cc
-#include <ostream>
-
-${namespace_open}
-
-${name}::${name}() : m_obj(new ${name}Obj()){
- m_obj->acquire();
-}
-
-${constructor_implementation}
-
-${name}::${name}(const ${name}& other) : m_obj(other.m_obj) {
-  m_obj->acquire();
-}
-
-${name}& ${name}::operator=(const ${name}& other) {
-  if (m_obj) m_obj->release();
-  m_obj = other.m_obj;
-  return *this;
-}
-
-${name}::${name}(${name}Obj* obj) : m_obj(obj){
-  if(m_obj)
-    m_obj->acquire();
-}
-
-${name} ${name}::clone() const {
-  return {new ${name}Obj(*m_obj)};
-}
-
-${name}::~${name}(){
-  if (m_obj) m_obj->release();
-}
-
-${name}::operator Const${name}() const {return Const${name}(m_obj);}
-
-$getters
-$setters
-$relations
-$extracode
-bool  ${name}::isAvailable() const {
-  if (m_obj) {
-    return true;
-  }
-  return false;
-}
-
-const podio::ObjectID ${name}::getObjectID() const {
-  if (m_obj) {
-    return m_obj->id;
-  }
-  return podio::ObjectID{-2,-2};
-}
-
-bool ${name}::operator==(const Const${name}& other) const {
-  return (m_obj == other.m_obj);
-}
-
-${ostream_implementation}
-
-//bool operator< (const ${name}& p1, const ${name}& p2 ) {
-//  if( p1.m_containerID == p2.m_containerID ) {
-//    return p1.m_index < p2.m_index;
-//  } else {
-//    return p1.m_containerID < p2.m_containerID;
-//  }
-//}
-
-${namespace_close}
