--- python/templates/Object.constructor.cc.template
@@ -1,4 +1,0 @@
-${name}::${name}(${signature}) : m_obj(new ${name}Obj()) {
-  m_obj->acquire();
-${constructor}
-}
