+++ python/templates/Object.h.jinja2
