--- python/templates/Object.h.template
@@ -1,89 +1,0 @@
-#ifndef ${PACKAGE_NAME}_${name}_H
-#define ${PACKAGE_NAME}_${name}_H
-
-#include "${incfolder}${name}Const.h"
-#include "${incfolder}${name}Obj.h"
-
-$includes
-
-#include "podio/ObjectID.h"
-
-#include <vector>
-#include <ostream>
-#include <iomanip>
-
-//forward declarations
-$forward_declarations
-
-${namespace_open}
-
-class ${name}Collection;
-class ${name}CollectionIterator;
-class Const${name};
-
-/** @class ${name}
- *  $description
- *  @author: $author
- */
-class ${name} {
-
-  friend ${name}Collection;
-  friend ${name}CollectionIterator;
-  friend Const${name};
-
-public:
-
-  /// default constructor
-  ${name}();
-${constructor_declaration}
-  /// constructor from existing ${name}Obj
-  ${name}(${name}Obj* obj);
-  /// copy constructor
-  ${name}(const ${name}& other);
-  /// copy-assignment operator
-  ${name}& operator=(const ${name}& other);
-  /// support cloning (deep-copy)
-  ${name} clone() const;
-  /// destructor
-  ~${name}();
-
-  /// conversion to const object
-  operator Const${name} () const;
-
-public:
-
-$getter_declarations
-$setter_declarations
-$relation_declarations
-$extracode_declarations
-  /// check whether the object is actually available
-  bool isAvailable() const;
-  /// disconnect from ${name}Obj instance
-  void unlink(){m_obj = nullptr;}
-
-  bool operator==(const ${name}& other) const {
-    return (m_obj == other.m_obj);
-  }
-
-  bool operator==(const Const${name}& other) const;
-
-// less comparison operator, so that objects can be e.g. stored in sets.
-//  friend bool operator< (const ${name}& p1,
-//       const ${name}& p2 );
-  bool operator<(const ${name}& other) const { return m_obj < other.m_obj; }
-
-
-  unsigned int id() const { return getObjectID().collectionID * 10000000 + getObjectID().index; }
-
-  const podio::ObjectID getObjectID() const;
-
-private:
-  ${name}Obj* m_obj;
-
-};
-
-${ostream_declaration}
-
-${namespace_close}
-
-#endif
