--- python/templates/RefVector.cc.template
@@ -1,35 +1,0 @@
-std::vector<${relationtype}>::const_iterator ${classname}::${relation}_begin() const {
-  auto ret_value = m_obj->m_${relation}->begin();
-  std::advance(ret_value, m_obj->data.${relation}_begin);
-  return ret_value;
-}
-
-std::vector<${relationtype}>::const_iterator ${classname}::${relation}_end() const {
-  auto ret_value = m_obj->m_${relation}->begin();
-  std::advance(ret_value, m_obj->data.${relation}_end);
-  return ret_value;
-}
-
-void ${classname}::${add_relation}(${relationtype} component) {
-  m_obj->m_${relation}->push_back(component);
-  m_obj->data.${relation}_end++;
-}
-
-unsigned int ${classname}::${relation}_size() const {
-  return (m_obj->data.${relation}_end - m_obj->data.${relation}_begin);
-}
-
-${relationtype} ${classname}::${get_relation}(unsigned int index) const {
-  if (${relation}_size() > index) {
-    return m_obj->m_${relation}->at(m_obj->data.${relation}_begin+index);
-  }
-  else throw std::out_of_range ("index out of bounds for existing references");
-}
-
-podio::RelationRange<${relationtype}> ${classname}::${get_relation}() const {
-  auto begin = m_obj->m_${relation}->begin();
-  std::advance(begin, m_obj->data.${relation}_begin);
-  auto end = m_obj->m_${relation}->begin();
-  std::advance(end, m_obj->data.${relation}_end);
-  return {begin, end};
-}
