+++ python/templates/macros/collections.jinja2
