+++ python/templates/macros/declarations.jinja2
