+++ python/templates/macros/implementations.jinja2
