+++ python/templates/selection.xml.jinja2
