--- python/templates/selection.xml.template
@@ -1,7 +1,0 @@
-<lcgdict>
-    <selection>
-
-$classes
-
-    </selection>
-</lcgdict>
